backend/requirements.txt
fastapi==0.85.1
uvicorn==0.18.3
SQLAlchemy==1.4.18
python-jose==3.3.0
passlib==1.7.1
python-dotenv==0.21.0
pymysql==1.0.2
sqlmodel==0.08

Faker==15.0.0